+++ src/main/resources/db/changelog/changes/003-create-sessions-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE sessions (
     id UUID PRIMARY KEY,
-    user_id int NOT NULL,
+    user_id bigint NOT NULL,
     expires_at TIMESTAMP WITH TIME ZONE NOT NULL,
 
     CONSTRAINT fk_sessions_user_id FOREIGN KEY (user_id) REFERENCES users(id)
